+++ lib/models/create_order.dart
@@ -1,4 +1,0 @@
-class CreateOrder 
-{
-  // TODO: Make create order model
-}